Add SchedulePlanningService.kt
