+++ tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/model/solver/PlanningSchedule.kt
@@ -1,11 +1,10 @@
-package ru.thetenzou.tsoddservice.solver
+package ru.thetenzou.tsoddservice.model.solver
 
 import org.optaplanner.core.api.domain.solution.PlanningEntityCollectionProperty
 import org.optaplanner.core.api.domain.solution.PlanningScore
 import org.optaplanner.core.api.domain.solution.PlanningSolution
 import org.optaplanner.core.api.domain.valuerange.ValueRangeProvider
 import org.optaplanner.core.api.score.buildin.hardsoft.HardSoftScore
-import ru.thetenzou.tsoddservice.model.schedule.ScheduledTask
 import java.time.LocalDate
 
 @PlanningSolution
